Add get_film_data script
It's only working with a single film so far, but this script scrapes IMDb to get information all about the film and then writes it out to CSV. The plan is for it to read in from the diva_film_ids CSV and pull information on all of those films.

Also: lint get_film_ids script + delete old script.